Minor reorganization of the state folder structure
