+++ web-app/src/index.js
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import ReactDOM from 'react-dom';
 import { App } from './components/stateful';
-import { appReducer } from './store';
+import { appReducer } from './store/reducers';
 import * as serviceWorker from './serviceWorker';
 import './index.css';
