+++ web-app/src/store/action-types/l10n/index.ts
@@ -1,0 +1,1 @@
+export * from './l10n.action-type';
