+++ web-app/src/store/action-types/l10n/l10n.action-type.ts
@@ -1,0 +1,1 @@
+export const L10N_SWITCH_LANGUAGE = "L10N_SWITCH_LANGUAGE";
