+++ web-app/src/store/actions/l10n/l10n.action.tsx
@@ -1,6 +1,5 @@
 import { SwitchLanguage } from '../../../types';
-
-export const L10N_SWITCH_LANGUAGE = "L10N_SWITCH_LANGUAGE";
+import { L10N_SWITCH_LANGUAGE } from '../../../store/action-types';
 
 export const switchLanguage: SwitchLanguage = (language) => ({
   type: L10N_SWITCH_LANGUAGE,
