--- web-app/src/store/index.ts
@@ -1,2 +1,0 @@
-export * from './actions';
-export * from './reducers';
