+++ web-app/src/store/reducers/app/app.reducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { AppState } from '../../../types';
-import { l10nReducer } from '../../reducers';
+import { l10nReducer } from '../../../store/reducers';
 
 export const appReducer = combineReducers<AppState>({
   l10n: l10nReducer
