+++ web-app/src/store/reducers/l10n/l10n.reducer.ts
@@ -2,7 +2,7 @@ import {
   L10nState,
   L10nActions,
 } from '../../../types';
-import { L10N_SWITCH_LANGUAGE } from '../../actions';
+import { L10N_SWITCH_LANGUAGE } from '../../../store/action-types';
 
 const l10nInitialState: L10nState = {
   language: "en"
