+++ web-app/src/types/l10n/switch-language-action.type.ts
@@ -2,6 +2,6 @@ import {
   Action,
   SwitchLanguagePayload
 } from '../../types';
-import { L10N_SWITCH_LANGUAGE } from '../../store/actions';
+import { L10N_SWITCH_LANGUAGE } from '../../store/action-types';
 
 export type SwitchLanguageAction = Action<typeof L10N_SWITCH_LANGUAGE, SwitchLanguagePayload>
